app/requirements.txt
# Database
pgvector~=0.2.5
psycopg[binary]~=3.2.3
peewee

# Web Scraping
newspaper4k[all]
lxml[html_clean]
playwright
beautifulsoup4>=4.12.0

# NLP & ML
spacy[apple]>=3.7.2
numpy>=1.22.0

# Utils
python-dotenv
ell-ai[studio,sqlite]~=0.0.15
tenacity

# Text to Speech
kokoro-onnx
soundfile

# AI
groq
openai
